+++ dflib/src/main/java/com/nhl/dflib/builder/ObjectAccum.java
@@ -27,6 +27,13 @@ public class ObjectAccum<T> implements ValueAccum<T> {
         return size;
     }
 
+    /**
+     * @since 1.0.0-M19
+     */
+    public void fill(Series<T> values, int valuesOffset, int accumOffset, int len) {
+        values.copyTo(data, valuesOffset, accumOffset, len);
+    }
+
     public void fill(int from, int to, T value) {
         if (to - from < 1) {
             return;