+++ dflib/src/main/java/com/nhl/dflib/series/IntBaseSeries.java
@@ -13,6 +13,7 @@ import com.nhl.dflib.ValueMapper;
 import com.nhl.dflib.ValuePredicate;
 import com.nhl.dflib.ValueToRowMapper;
 import com.nhl.dflib.builder.IntAccum;
+import com.nhl.dflib.builder.ObjectAccum;
 import com.nhl.dflib.builder.UniqueIntAccum;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.groupby.SeriesGrouper;
@@ -373,6 +374,47 @@ public abstract class IntBaseSeries implements IntSeries {
         }
 
         return new BooleanArraySeries(data);
+    }
+
+    @Override
+    public Series<Integer> replace(IntSeries positions, Series<Integer> with) {
+        int rs = positions.size();
+        if (rs != with.size()) {
+            throw new IllegalArgumentException("Positions size " + rs + " is not the same replacement Series size " + with.size());
+        }
+
+        if (rs == 0) {
+            return this;
+        }
+
+        int s = size();
+
+        // Quick check for nulls in "with". May result in false positives (no nulls in Series<Integer), but does not
+        // require checking each value
+
+        if (with instanceof IntSeries) {
+
+            IntSeries withInt = (IntSeries) with;
+            IntAccum values = new IntAccum(s);
+
+            values.fill(this, 0, 0, s);
+
+            for (int i = 0; i < rs; i++) {
+                values.replace(positions.getInt(i), withInt.getInt(i));
+            }
+
+            return values.toSeries();
+        } else {
+            ObjectAccum<Integer> values = new ObjectAccum<>(s);
+
+            values.fill(this, 0, 0, s);
+
+            for (int i = 0; i < rs; i++) {
+                values.replace(positions.getInt(i), with.get(i));
+            }
+
+            return values.toSeries();
+        }
     }
 
     @Override