+++ dflib/src/test/java/com/nhl/dflib/LongSeries_ReplaceTest.java
@@ -8,6 +8,26 @@ import static org.junit.jupiter.api.Assertions.*;
 public class LongSeries_ReplaceTest {
 
     @Test
+    public void replace_positions() {
+
+        Series<Long> s1 = Series.ofLong(1, 0, 2, -1).replace(
+                Series.ofInt(1, 3),
+                Series.ofLong(10, 100));
+
+        new SeriesAsserts(s1).expectData(1L, 10L, 2L, 100L);
+    }
+
+    @Test
+    public void replace_positions_nulls() {
+
+        Series<Long> s1 = Series.ofLong(1, 0, 2, -1).replace(
+                Series.ofInt(1, 3),
+                Series.of(10L, null));
+
+        new SeriesAsserts(s1).expectData(1L, 10L, 2L, null);
+    }
+
+    @Test
     public void replace() {
         BooleanSeries cond = Series.ofBool(true, true, false, false);